chore: migrate test stack to @cofhe/sdk + redeploy contracts
- Bump @fhenixprotocol/cofhe-contracts 0.0.13 to 0.1.4
- Swap cofhejs + cofhe-hardhat-plugin + @fhenixprotocol/cofhe-mock-contracts
  for @cofhe/sdk@0.6.0 + @cofhe/hardhat-plugin@0.6.0 +
  @cofhe/mock-contracts@0.6.0
- Rewrite test helpers on createClientWithBatteries + encryptInputs +
  decryptForView; 49/49 passing on hardhat mocks
- Cast bytes32 to uint256 at OrderQuoted / OrderInProccessed / OrderInQueued
  / UnwrapRequested event emits so the event ABI stays stable for frontend
  + observer decoders against cofhe-contracts 0.1.4
- Redeploy Sigill + cUSDC on Base Sepolia; live smoke-quote tx confirms
  @cofhe/sdk@0.6.0 + cofhe-contracts@0.1.4 interop end-to-end
- Update README, docs/wave5.md, landing constants with new addresses
- Drop obsolete e2e-cusdc.ts (cofhejs-era); add scripts/smoke-quote.ts

Sigill: 0xb302566eA948f2039Cf6f8a30719F56a93e49Ab4
cUSDC:  0x285b239fc9fE8B100d0Cb865cb44BdB166f81977

+++ packages/contracts/contracts/ConfidentialERC20.sol
@@ -22,7 +22,7 @@ contract ConfidentialERC20 {
     ///         `decryptForTx`) that hasn't shipped to npm yet. Until that
     ///         arrives we delegate the unseal to a permissioned operator
     ///         who already has `FHE.allow(debit, unwrapper)` and produces
-    ///         plaintexts off-chain via `cofhejs.unseal`.
+    ///         plaintexts off-chain via `client.decryptForView`.
     address public immutable unwrapper;
     string public name;
     string public symbol;
@@ -88,7 +88,7 @@ contract ConfidentialERC20 {
         _debit(msg.sender, debit);
 
         // ACL so the contract + recipient + unwrapper can all reference the
-        // debit handle. The unwrapper uses its ACL to call `cofhejs.unseal`
+        // debit handle. The unwrapper uses its ACL to call `client.decryptForView`
         // and get the plaintext; nobody else can read the value.
         FHE.allowThis(debit);
         FHE.allow(debit, msg.sender);
@@ -98,12 +98,12 @@ contract ConfidentialERC20 {
         pendingUnwraps[unwrapId] = PendingUnwrap({recipient: msg.sender, encAmount: debit, claimed: false});
         // Emit the raw handle so off-chain listeners can unseal without a
         // second read against `pendingUnwraps`.
-        emit UnwrapRequested(unwrapId, msg.sender, euint64.unwrap(debit));
+        emit UnwrapRequested(unwrapId, msg.sender, uint256(euint64.unwrap(debit)));
     }
 
     /// @notice Finalise an unwrap. Either the trusted `unwrapper` or the
     ///         original `recipient` may submit `plain`. Both paths are
-    ///         trusted to unseal the debit handle off-chain via cofhejs and
+    ///         trusted to unseal the debit handle off-chain via `cofhe-sdk` and
     ///         pass the honest value; the recipient has ACL via
     ///         `FHE.allow(debit, msg.sender)` set in `requestUnwrap`.
     ///         Demo trust model until Fhenix's SDK `decryptForTx` ships, at
@@ -163,7 +163,7 @@ contract ConfidentialERC20 {
         FHE.allow(transferred, msg.sender);
     }
 
-    /// @notice EOA-facing approve: caller signs an `InEuint64` (via cofhejs)
+    /// @notice EOA-facing approve: caller signs an `InEuint64` (via `cofhe-sdk`)
     ///         for the amount they want to authorize. `FHE.asEuint64` mints a
     ///         fresh handle owned by the caller, so subsequent `FHE.allow*`
     ///         calls succeed.
@@ -176,7 +176,7 @@ contract ConfidentialERC20 {
     /// @notice Contract-facing approve: caller passes an existing `euint64`
     ///         handle that it already has manage rights on. Useful when an
     ///         escrow contract holds its own funds and wants to authorize a
-    ///         spender without round-tripping through cofhejs.
+    ///         spender without round-tripping through `cofhe-sdk`.
     function approve(address spender, euint64 amount) external returns (euint64) {
         _approve(msg.sender, spender, amount);
         return amount;
